Quik ClientLevelObject fix

+++ gdx-bunny-brawl/src/main/java/de/hochschuletrier/gdw/ws1314/entity/levelObjects/client/ClientLevelObject.java
@@ -37,7 +37,7 @@ public abstract class ClientLevelObject extends ClientEntity
 	}
 
 	@Override
-	public void update(int delta)
+	public void update(float delta)
 	{
 	}
